pyproject.toml
[tool.poetry]
name = "rabbitmq-python-producer-consumer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.11.10"
aio-pika = "^9.5.3"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
